+++ minecraft-discord-bridge/__main__.py
@@ -1,0 +1,3 @@
+if __name__ == "__main__":
+    from . import minecraft_discord_bridge
+    minecraft_discord_bridge.main()
