+++ minecraft-discord-bridge/database.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta, timezone
 from sqlalchemy import Column, String, Integer, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 
-from database_session import Base
+from .database_session import Base
 
 
 class DiscordChannel(Base):
