+++ minecraft-discord-bridge/elasticsearch_logger.py
